requirements.txt
pymongo
python-dotenv
pydantic
requests